+++ rake_build/combine_sources.rb
