+++ rake_build/generate_stats.rb
