+++ rake_build/verify_source_data.rb
